@@ -1,116 +1,116 @@
 {
-    "name": "bottleneck",
-    "version": "2.0.0",
-    "engines": {
-        "node": ">= 7.6.0"
-    },
-    "description": "Multiplayer RPG with scripting-based problem solving",
-    "main": "main.js",
-    "dependencies": {
-        "@fortawesome/fontawesome-free": "^5.1.0",
-        "ace-builds": "^1.3.3",
-        "ajv": "^6.5.2",
-        "awesome-typescript-loader": "^5.2.0",
-        "babel-core": "^6.25.0",
-        "babel-loader": "^7.1.2",
-        "babel-plugin-module-resolver": "^2.7.1",
-        "babel-preset-env": "^1.7.0",
-        "babel-preset-react": "^6.24.1",
-        "babel-preset-stage-3": "^6.24.1",
-        "bcrypt-nodejs": "0.0.3",
-        "bootstrap": "^4.1.2",
-        "bootstrap-sass": "^3.3.7",
-        "cerialize": "^0.1.18",
-        "dotenv": "^6.0.0",
-        "express": "^4.14.0",
-        "jimp": "github:oliver-moran/jimp#master",
-        "jquery": "^3.3.1",
-        "js-graph-algorithms": "^1.0.18",
-        "lance-gg": "github:lolney/lance#master",
-        "moment": "^2.22.2",
-        "octicons": "^7.4.0",
-        "paper": "^0.11.5",
-        "pathfinding": "^0.4.18",
-        "pg": "^7.4.3",
-        "pixi.js": "^4.7.1",
-        "popper": "^1.0.1",
-        "prop-types": "^15.6.1",
-        "query-string": "^4.2.3",
-        "react": "^16.4.0",
-        "react-ace": "^6.1.1",
-        "react-bootstrap": "^0.32.1",
-        "react-dom": "^16.4.0",
-        "react-draggable": "^3.0.5",
-        "react-modal": "^3.4.4",
-        "sequelize": "^4.37.10",
-        "socket.io": "^1.4.8",
-        "socketio-auth": "^0.1.1",
-        "source-map-loader": "^0.2.3",
-        "style-loader": "^0.21.0",
-        "typescript": "^2.9.2",
-        "uuid": "^3.2.1"
-    },
-    "devDependencies": {
-        "@kadira/storybook-addons": "^1.6.1",
-        "@storybook/addon-actions": "^4.0.0-alpha.3",
-        "@storybook/addon-links": "^4.0.0-alpha.3",
-        "@storybook/addons": "^4.0.0-alpha.3",
-        "@storybook/react": "^4.0.0-alpha.3",
-        "babel-cli": "^6.26.0",
-        "babel-polyfill": "^6.26.0",
-        "babel-register": "^6.26.0",
-        "babel-runtime": "^6.26.0",
-        "babelify": "^7.3.0",
-        "buffer": "^5.1.0",
-        "css-loader": "^0.28.11",
-        "enzyme": "^3.3.0",
-        "enzyme-adapter-react-16": "^1.1.1",
-        "eslint": "^5.1.0",
-        "eslint-config-google": "^0.7.1",
-        "eslint-plugin-jasmine": "^2.10.1",
-        "eslint-plugin-newline-after-if-condition": "^1.1.0",
-        "eslint-plugin-react": "^7.10.0",
-        "file-loader": "^1.1.11",
-        "jasmine": "^3.1.0",
-        "jsdom": "^11.11.0",
-        "node-sass": "^4.9.0",
-        "raw-loader": "^0.5.1",
-        "react-jsdom": "git+https://github.com/tomcardoso/react-jsdom.git",
-        "react-storybook-console": "^2.0.4",
-        "sass-loader": "^7.0.1",
-        "sequelize-auto-migrations": "^1.0.2",
-        "sequelize-cli": "^4.0.0",
-        "sinon": "^5.0.10",
-        "speed-measure-webpack-plugin": "^1.2.2",
-        "string-replace-loader": "^2.1.1",
-        "webpack": "^4.14.0",
-        "webpack-cli": "^3.0.0",
-        "webpack-merge": "^4.1.3"
-    },
-    "scripts": {
-        "start": "babel-node main.js --presets env",
-        "test": "babel-node spec/run.js --presets env",
-        "start-debug": "babel-node main.js --inspect --debug-brk --presets env",
-        "start-dev": "babel-watch main.js --presets env",
-        "nodemon": "nodemon main.js --exec babel-node --presets env",
-        "postinstall": "webpack --config webpack.prod.js --display-error-details && bash migrations.sh",
-        "build": "webpack --config webpack.prod.js",
-        "dev": "webpack --watch --config webpack.dev.js",
-        "stats": "webpack --env production --profile --json > stats.json",
-        "storybook": "start-storybook -p 9001 -s . -c  .storybook",
-        "build-storybook": "build-storybook"
-    },
-    "repository": {
-        "type": "git",
-        "url": ""
-    },
-    "keywords": [
-        "game",
-        "lance",
-        "lance-gg",
-        "multiplayer",
-        "websockets"
-    ],
-    "author": "Luke Olney",
-    "license": "Apache-2.0"
+  "name": "bottleneck",
+  "version": "2.0.0",
+  "engines": {
+    "node": "8.10.0"
+  },
+  "description": "Multiplayer RPG with scripting-based problem solving",
+  "main": "main.js",
+  "dependencies": {
+    "@fortawesome/fontawesome-free": "^5.1.0",
+    "ace-builds": "^1.3.3",
+    "ajv": "^6.5.2",
+    "awesome-typescript-loader": "^5.2.0",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-module-resolver": "^2.7.1",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-3": "^6.24.1",
+    "bcrypt-nodejs": "0.0.3",
+    "bootstrap": "^4.1.2",
+    "bootstrap-sass": "^3.3.7",
+    "cerialize": "^0.1.18",
+    "dotenv": "^6.0.0",
+    "express": "^4.14.0",
+    "jimp": "^0.5.3",
+    "jquery": "^3.3.1",
+    "js-graph-algorithms": "^1.0.18",
+    "lance-gg": "github:lolney/lance#master",
+    "moment": "^2.22.2",
+    "octicons": "^7.4.0",
+    "paper": "^0.11.5",
+    "pathfinding": "^0.4.18",
+    "pg": "^7.4.3",
+    "pixi.js": "^4.7.1",
+    "popper": "^1.0.1",
+    "prop-types": "^15.6.1",
+    "query-string": "^4.2.3",
+    "react": "^16.4.0",
+    "react-ace": "^6.1.1",
+    "react-bootstrap": "^0.32.1",
+    "react-dom": "^16.4.0",
+    "react-draggable": "^3.0.5",
+    "react-modal": "^3.4.4",
+    "sequelize": "^4.37.10",
+    "socket.io": "^1.4.8",
+    "socketio-auth": "^0.1.1",
+    "source-map-loader": "^0.2.3",
+    "style-loader": "^0.21.0",
+    "typescript": "^2.9.2",
+    "uuid": "^3.2.1"
+  },
+  "devDependencies": {
+    "@kadira/storybook-addons": "^1.6.1",
+    "@storybook/addon-actions": "^4.0.0-alpha.3",
+    "@storybook/addon-links": "^4.0.0-alpha.3",
+    "@storybook/addons": "^4.0.0-alpha.3",
+    "@storybook/react": "^4.0.0-alpha.3",
+    "babel-cli": "^6.26.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-register": "^6.26.0",
+    "babel-runtime": "^6.26.0",
+    "babelify": "^7.3.0",
+    "buffer": "^5.1.0",
+    "css-loader": "^0.28.11",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "eslint": "^5.1.0",
+    "eslint-config-google": "^0.7.1",
+    "eslint-plugin-jasmine": "^2.10.1",
+    "eslint-plugin-newline-after-if-condition": "^1.1.0",
+    "eslint-plugin-react": "^7.10.0",
+    "file-loader": "^1.1.11",
+    "jasmine": "^3.1.0",
+    "jsdom": "^11.11.0",
+    "node-sass": "^4.9.0",
+    "raw-loader": "^0.5.1",
+    "react-jsdom": "git+https://github.com/tomcardoso/react-jsdom.git",
+    "react-storybook-console": "^2.0.4",
+    "sass-loader": "^7.0.1",
+    "sequelize-auto-migrations": "^1.0.2",
+    "sequelize-cli": "^4.0.0",
+    "sinon": "^5.0.10",
+    "speed-measure-webpack-plugin": "^1.2.2",
+    "string-replace-loader": "^2.1.1",
+    "webpack": "^4.14.0",
+    "webpack-cli": "^3.0.0",
+    "webpack-merge": "^4.1.3"
+  },
+  "scripts": {
+    "start": "babel-node main.js --presets env",
+    "test": "babel-node spec/run.js --presets env",
+    "start-debug": "babel-node main.js --inspect --debug-brk --presets env",
+    "start-dev": "babel-watch main.js --presets env",
+    "nodemon": "nodemon main.js --exec babel-node --presets env",
+    "postinstall": "webpack --config webpack.prod.js --display-error-details && bash migrations.sh",
+    "build": "webpack --config webpack.prod.js",
+    "dev": "webpack --watch --config webpack.dev.js",
+    "stats": "webpack --env production --profile --json > stats.json",
+    "storybook": "start-storybook -p 9001 -s . -c  .storybook",
+    "build-storybook": "build-storybook"
+  },
+  "repository": {
+    "type": "git",
+    "url": ""
+  },
+  "keywords": [
+    "game",
+    "lance",
+    "lance-gg",
+    "multiplayer",
+    "websockets"
+  ],
+  "author": "Luke Olney",
+  "license": "Apache-2.0"
 }
