@@ -25,6 +25,19 @@ export default class GameWorld {
         this.pathFinder = new PF.AStarFinder();
     }
 
+    static getResourceBounds() {
+        let gameBounds = Bounds.fromDimensions(WIDTH, HEIGHT);
+        let halfBounds = gameBounds.scale(0.48, 1);
+
+        let { left, top, bottom, center: mazeBounds } = halfBounds.crop(
+            0.25,
+            1,
+            0.3,
+            0.7
+        );
+        return top;
+    }
+
     static generate() {
         let gameBounds = Bounds.fromDimensions(WIDTH, HEIGHT);
         let halfBounds = gameBounds.scale(0.48, 1);
@@ -86,12 +99,13 @@ export default class GameWorld {
         let start = this.grid.worldCoordsToCell(mapStart);
         let end = this.grid.worldCoordsToCell(mapEnd);
 
+        console.log('grid start x,y, end x,y', start.x, start.y, end.x, end.y);
         if (this.grid.isOccupied(end)) {
             throw new Error('end tile is unreachable');
         }
 
         let grid = new PF.Grid(this.grid.to2DArray());
-        console.log('grid start x,y, end x,y', start.x, start.y, end.x, end.y);
+
         let path = this.pathFinder.findPath(
             start.x,
             start.y,
@@ -193,8 +207,8 @@ export class Grid {
         let x = worldCoords.x % WIDTH;
         let y = worldCoords.y % HEIGHT;
         return new TwoVector(
-            Math.round(this.resolution * (x / WIDTH)),
-            Math.round(this.resolution * (y / HEIGHT))
+            Math.floor(this.resolution * (x / WIDTH)),
+            Math.floor(this.resolution * (y / HEIGHT))
         );
     }
 
@@ -491,7 +505,7 @@ export class MazeWall {
         let isHorizonal = width == 1;
 
         this.width = isHorizonal ? scale : maze.wallWidth;
-        this.height = isHorizonal ? maze.wallWidth : maze.corridorWidth;
+        this.height = isHorizonal ? maze.wallWidth : scale + maze.wallWidth;
 
         let xPadding = 0;
         let yPadding = isHorizonal
