@@ -37,19 +37,34 @@ export default class MyServerEngine extends ServerEngine {
 
     onPlayerConnected(socket) {
         super.onPlayerConnected(socket);
-        Controller.addPlayer(socket.playerId, socket);
-        this.gameEngine.makePlayer(socket.playerId);
+        let waitForAuth = () => {
+            if (socket.auth) {
+                console.log('Authenticated. Creating player.');
+                Controller.addPlayer(
+                    socket.client.playerDbId,
+                    socket.playerId,
+                    socket
+                );
+                this.gameEngine.makePlayer(
+                    socket.client.playerDbId,
+                    socket.playerId
+                );
+            } else setTimeout(waitForAuth, 100);
+        };
+        waitForAuth();
     }
 
-    onPlayerDisconnected(socketId, playerId) {
-        super.onPlayerDisconnected(socketId, playerId);
-        console.log(`removing player ${playerId}`);
-        let playerObjects = this.gameEngine.world.queryObjects({
-            playerId: playerId
+    onPlayerDisconnected(socketId, playerNumber) {
+        super.onPlayerDisconnected(socketId, playerNumber);
+        console.log(`removing player ${playerNumber}`);
+        let playerObjects = this.gameEngine.queryObjects({
+            playerNumber: playerNumber
         });
         playerObjects.forEach((obj) => {
             this.gameEngine.removeObjectFromWorld(obj.id);
+            if (obj.playerId) {
+                Controller.removePlayer(obj.playerId);
+            }
         });
-        Controller.removePlayer(playerId);
     }
 }
